refactor: ComputerController를 ComputerPlayer로 변경
